SWBF lua uses float instead of double for lua_Number. Probably because PS2 doesn't support doubles.

+++ include/lua.h
@@ -85,7 +85,8 @@ typedef int (*lua_Chunkwriter) (lua_State *L, const void* p,
 
 /* type of numbers in Lua */
 #ifndef LUA_NUMBER
-typedef double lua_Number;
+//typedef double lua_Number;
+typedef float lua_Number;	// in swbf2 they used float instead of double
 #else
 typedef LUA_NUMBER lua_Number;
 #endif
